requirements.txt
autopep8==2.0.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
Flask==2.2.3
gunicorn==20.1.0
fonttools==4.38.0
google-api-core==2.11.0
google-api-python-client==2.85.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-cloud-aiplatform==1.25.0
google-cloud-bigquery==3.10.0
google-cloud-core==2.3.2
google-cloud-resource-manager==1.10.1
google-cloud-storage==2.9.0
google-cloud-vision==3.4.1
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.53.0
grpcio-status==1.53.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.5.3
numpy==1.21.6
opencv-python==4.7.0.72
packaging==23.1
pandas==1.3.5
Pillow==9.5.0
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3
requests==2.28.2
rsa==4.9
Shapely==1.8.5.post1
six==1.16.0
tomli==2.0.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
Werkzeug==2.2.3
